requirements.txt
asgiref==3.3.4
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-crispy-forms==1.12.0
django-dotenv==1.4.2
django-tinymce4-lite==1.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsmin==2.2.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==8.2.0
platformdirs==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
tomli==1.2.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
